refactor(utils): extract notification helpers to shared utils
