+++ src/utils/formatters.js
@@ -60,6 +60,17 @@ export function formatAvgPrice(value, numberFormat) {
   }
 }
 
+export function getTimeAgo(timestamp) {
+  const seconds = Math.floor((Date.now() - timestamp) / 1000);
+  if (seconds < 60) return 'just now';
+  const minutes = Math.floor(seconds / 60);
+  if (minutes < 60) return `${minutes}m ago`;
+  const hours = Math.floor(minutes / 60);
+  if (hours < 24) return `${hours}h ago`;
+  const days = Math.floor(hours / 24);
+  return `${days}d ago`;
+}
+
 export const formatTimer = (endTime) => {
   if (!endTime) return '--:--:--';
   const remaining = endTime - Date.now();
